requirements.txt
# ============================================================
# FinSight AI — Python Dependencies
# Install: pip install -r requirements.txt
# ============================================================

# --- Core API Framework ---
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20

# --- Configuration ---
pydantic>=2.10.0
pydantic-settings>=2.7.0

# --- Security ---
cryptography>=44.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# --- Document Text Extraction ---
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

# --- Local Embeddings (all-MiniLM-L6-v2, 384-dim, ~80 MB download) ---
sentence-transformers>=3.3.0

# --- Vector Store ---
chromadb>=0.6.0

# --- Text Chunking ---
langchain-text-splitters>=0.3.0

# --- LLM: Ollama Python library (local llama3.1 — native SDK, no raw HTTP) ---
ollama>=0.4.0

# --- LLM: Gemini (optional cloud fallback) ---
google-generativeai>=0.8.0

# --- Frontend (React/Vite — see frontend/package.json) ---
# requests kept only if needed for backend testing or scripting
requests>=2.32.0

# --- Logging ---
python-json-logger>=3.0.0

# --- Testing ---
pytest>=8.0.0
pytest-asyncio>=0.25.0
